Ajout de Route TD6
