+++ TD-React/package.json
@@ -15,8 +15,10 @@
     "@mui/icons-material": "^6.3.0",
     "@mui/material": "^6.3.0",
     "@mui/styled-engine-sc": "^6.3.0",
+    "file-saver": "^2.0.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.1",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
@@ -29,6 +31,6 @@
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.16",
     "globals": "^15.13.0",
-    "vite": "^6.0.5"
+    "vite": "^6.0.6"
   }
 }
